Rename npmPluginLoader to pluginNpmLoader for consistency
